# internal/gateway/platform-openapi.yaml
openapi: 3.0.3
info:
  title: Kiban platform API
  version: "1.0.0"
  description: >
    The gateway-owned "platform" surface: every `/api/*` route mounted directly by
    `internal/gateway` itself (foundation_routes.go, platform_routes.go, admin_position_routes.go,
    admin_group_routes.go), as opposed to a module's own `/api/<moduleKey>/*` fragment.
    This is not a module: there is no module.manifest.json/
    authz.fragment.json for the platform; it is documented here because the module contract's
    binding rule ("the running service must enforce this contract at the HTTP boundary") applies
    equally to the platform's own consumer-facing surface, so a developer can use Kiban from
    docs alone.

    Every route here is a fixed reverse proxy (never the dynamic, opaque module-key proxy in
    proxy.go): the gateway rewrites the outbound path to an internal `/internal/...` route on
    one of the four foundation services (authz, org, registry) and, for the self-scoped reads,
    forces the outbound `kcSub` to the validated bearer's own subject (never a client-supplied
    value: see foundation_routes.go's own doc comment). All routes require a valid bearer
    (`RequireAuth`); routes under `.../admin/...` additionally require the superadmin guard
    (`RequireSuperadmin`): a fail-closed check: if authz itself is unreachable, the admin
    guard returns 503 `AUTHORIZATION_UNAVAILABLE`, never a silent allow.

...
    document is trustworthy, not aspirational.
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/effective-access/can:
    post:
      operationId: effectiveAccessCan
      summary: Ask whether the caller (never an arbitrary actor) can do one thing
      description: >
        Bearer-gated fixed proxy to authz's own `/internal/authz/effective-access/can`
        (internal/authz/http.go). authz itself rejects a body naming a subject other than the
        bearer (400), and a `global`-scope request without `requiredPlatformRole` (400
        `VALIDATION_ERROR`), the gateway adds nothing beyond bearer validation at the edge.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EffectiveAccessRequest" }
      responses:
        "200":
          description: Decision (allow or deny, a denial is a normal 200, never an error status)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EffectiveAccessDecisionEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/AuthorizationUnavailable" }
  /api/auth/effective-access/batch-can:
    post:
      operationId: effectiveAccessBatchCan
      summary: Ask whether the caller can do several things in one round trip
      description: >
        Same subject and scope rules as `can`, bearer-gated fixed proxy to authz's own
        `/internal/authz/effective-access/batch-can`. At most 100 `items` per request (400
        `VALIDATION_ERROR` above); a request with denied items writes one audit event listing
        them (the first 50, with a `truncated` flag).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EffectiveAccessBatchRequest" }
      responses:
        "200":
          description: One decision per requested (object, relation) pair
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EffectiveAccessBatchEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/AuthorizationUnavailable" }
  /api/auth/effective-access/summary:
    get:
      operationId: effectiveAccessSummary
      summary: The caller's own display-composed access summary (a display aid, never an enforcement surface)
      description: >
        Self-scoped, internal-network-unauthenticated downstream (authz's own
        `/internal/authz/effective-access/summary`, kcSub as a plain query param), the gateway is
        the bearer-validating boundary and forces `kcSub` to the validated bearer's own subject,
        overwriting any client-supplied value. `objectAccess` is direct-tuples-only (never a
        userset-closure expansion);
        a position/group-held tier never appears here even when it is real and enforced elsewhere.
      parameters:
        - name: companyId
          in: query
          required: false
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: The caller's summary (apiVersion 1)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EffectiveAccessSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/org/me/companies:
    get:
      operationId: orgMeCompanies
      summary: Companies where the caller has an active membership in an active company (the company switcher's source)
      description: Self-scoped fixed proxy to org's own `/internal/org/me/companies` (internal/org/http.go's handleMeCompanies), kcSub forced to the bearer's own subject.
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/OrgMeCompany" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/org/companies/{companyId}/members:
    get:
      operationId: orgCompanyMemberDirectory
      summary: The company's member directory (the least-disclosure share-with/assign-to picker source)
      description: >
        Bearer-gated fixed proxy to org's own `/internal/org/companies/{companyId}/members`,
        kcSub forced to the bearer's own subject. org's own handler enforces
        active-membership-or-superadmin (`requireMemberOrAdmin`), a non-member caller gets
        403, not a 404 (never leaks whether the company exists to an outsider vs. a non-member).
      parameters:
        - name: companyId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: q
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
      responses:
        "200":
          description: Page of member-directory entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/MemberDirectoryEntry" }
                      total: { type: integer }
                      page: { type: integer }
                      pageSize: { type: integer }
                      totalPages: { type: integer }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/org/companies/{companyId}/members/by-subject/{subject}:
    get:
      operationId: orgMemberBySubject
      summary: Membership fact for one subject (app backend or superadmin)
      description: >
        Fixed proxy to org's `/internal/org/companies/{companyId}/members/by-kcsub/{subject}`.
        A registered app's backend (client-credentials token whose `azp` is the app's service
        client id) or a superadmin asks whether `subject` is an active member of the company and
        which member record that is. Never a directory: one subject, one answer.
      parameters:
        - name: companyId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: subject
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Membership fact
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      isMember: { type: boolean }
                      isActive: { type: boolean }
                      memberId: { type: string, format: uuid, nullable: true }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/AuthorizationUnavailable" }
  /api/auth/grants:
    post:
      operationId: authzGrants
      summary: Write authz tuples (superadmin, or a registered app on its own types)
      description: >
        Fixed proxy to authz's own `/internal/authz/grants` (internal/authz/http.go), the one
        write path exposed through the gateway for raw tuple grants. A superadmin may call it;
        so may a registered app's backend (a client-credentials token whose `azp` is the app's
        service client id), which authz then restricts to tuples on that app's own object types.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthzGrantsRequest" }
      responses:
        "200":
          description: Grant result
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      status: { type: string }
                      count: { type: integer }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422":
          description: A tuple names a base or unknown object type, or its object is not anchored to companyId (VALIDATION_FAILED)
        "503": { $ref: "#/components/responses/AuthorizationUnavailable" }
  /api/platform/admin/platform-roles:
    post:
      operationId: superadminGrantPlatformRole
      summary: Grant the platform role (superadmin only)
      description: >
        Superadmin-gated fixed proxy to authz's own `POST /internal/authz/platform-roles`. The
        platform role's one record is the authz tuple `system:platform#superadmin @ user:<subjectId>`;
        authz writes it with its grant-ledger and audit rows in one transaction. The subject must
        already exist in identity (404 otherwise); `kiban-superadmin` is the only role (422 otherwise).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlatformRoleRequest" }
      responses:
        "200":
          description: Granted (idempotent for an existing holder)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/PlatformRoleRequest" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "503": { $ref: "#/components/responses/AuthorizationUnavailable" }
  /api/platform/admin/platform-roles/{role}/{subjectId}:
    delete:
      operationId: superadminRevokePlatformRole
      summary: Revoke the platform role (superadmin only; the last superadmin cannot be revoked)
      description: >
        Superadmin-gated fixed proxy to authz's own `DELETE /internal/authz/platform-roles/{role}/{subjectId}`.
        404 when the subject does not hold the role; 409 CONFLICT when it is the platform's last
        superadmin (the caller revoking themselves included).
      parameters:
        - name: role
          in: path
          required: true
          schema: { type: string, enum: [kiban-superadmin] }
        - name: subjectId
          in: path
          required: true
          description: The subject's Keycloak subject id (kcSub).
          schema: { type: string }
      responses:
        "200":
          description: Revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/PlatformRoleRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "503": { $ref: "#/components/responses/AuthorizationUnavailable" }

  # ---- platform_routes.go ----------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        A Kiban access token: a user's, from a browser login through the gateway's /auth proxy
        (client kiban-frontend, PKCE), or a service client's, from the client-credentials grant
        (KIBAN_SERVICE_CLIENTS). Audience kiban-api, issuer <gateway>/realms/kiban.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
    PlatformRoleRequest:
      type: object
      required: [subjectId, role]
      properties:
        subjectId: { type: string, description: The subject's Keycloak subject id (kcSub). }
        role: { type: string, enum: [kiban-superadmin] }
    EffectiveAccessObjectRef:
      type: object
      properties:
        type: { type: string }
        id: { type: string }
    EffectiveAccessRequest:
      type: object
      required: [featureKey, scope]
      properties:
        featureKey: { type: string }
        moduleKey: { type: string }
        scope: { type: string, enum: [global, company] }
        companyId: { type: string, format: uuid }
        requiredPlatformRole: { type: string }
        allowPlatformOperatorCompanyScope: { type: boolean }
        requiredCompanyRole: { type: string }
        object: { $ref: "#/components/schemas/EffectiveAccessObjectRef" }
        relation: { type: string }
        requiresEligibility: { type: boolean }
        correlationId: { type: string }
    EffectiveAccessEvidence:
      type: object
      properties:
        key: { type: string }
        value: { type: string }
    EffectiveAccessDecision:
      type: object
      properties:
        allowed: { type: boolean }
        reason: { type: string }
        evidence:
          type: array
          items: { $ref: "#/components/schemas/EffectiveAccessEvidence" }
    EffectiveAccessDecisionEnvelope:
      type: object
      properties:
        data: { $ref: "#/components/schemas/EffectiveAccessDecision" }
    EffectiveAccessBatchRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          maxItems: 100
          items:
            type: object
            required: [object, relation]
            properties:
              object: { $ref: "#/components/schemas/EffectiveAccessObjectRef" }
              relation: { type: string }
    EffectiveAccessBatchEnvelope:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              object: { $ref: "#/components/schemas/EffectiveAccessObjectRef" }
              relation: { type: string }
              decision: { $ref: "#/components/schemas/EffectiveAccessDecision" }
    EffectiveAccessRoleBinding:
      type: object
      properties:
        scope: { type: string, enum: [global, company] }
        role: { type: string }
        companyId: { type: string, format: uuid }
    EffectiveAccessObjectAccessEntry:
      type: object
      properties:
        objectType: { type: string }
        objectId: { type: string }
        relations:
          type: array
          items: { type: string }
    EffectiveAccessSummary:
      type: object
      properties:
        data:
          type: object
          properties:
            apiVersion: { type: integer }
            subjectId: { type: string }
            companyId: { type: string }
            moduleKey: { type: string }
            featureKeys:
              type: array
              items: { type: string }
            roleBindings:
              type: array
              items: { $ref: "#/components/schemas/EffectiveAccessRoleBinding" }
            objectAccess:
              type: array
              items: { $ref: "#/components/schemas/EffectiveAccessObjectAccessEntry" }
            rowScopes: { type: array, items: {} }
            fieldPolicies: { type: array, items: {} }
...
    AuthzGrantsRequest:
      type: object
      required: [op, companyId, tuples]
      description: >
        Every tuple is bound to `companyId`: its object must be an object type declared by one
        enabled module's fragment (anchored to `company_module:<companyId>/<module>`, an anchor
        tuple in the same grant request counts) or `company_module:<companyId>/<module>` itself;
        base types (company, system, module, member, position, group) are refused (422). The
        bearer must pass the effective-access decision's steps 1–6 for that (module, company)
        (403 otherwise) and, for a `company_module` object, hold its `admin` relation.
      properties:
        op: { type: string, enum: [grant, revoke] }
        companyId: { type: string, format: uuid }
        correlationId: { type: string }
        tuples:
          type: array
          items:
            type: object
            required: [objectType, objectId, relation, subjectType, subjectId]
            properties:
              objectType: { type: string }
              objectId: { type: string }
              relation: { type: string }
              subjectType: { type: string }
              subjectId: { type: string }
              subjectRelation: { type: string }
    OrgMeCompany:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        isActive: { type: boolean }
    MemberDirectoryEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        displayName: { type: string }
        email: { type: string }
        hasLinkedUser: { type: boolean }
...
  responses:
    BadRequest:
      description: BAD_REQUEST / VALIDATION_ERROR
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: AUTH_TOKEN_MISSING / AUTH_TOKEN_INVALID
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: FORBIDDEN / AUTHORIZATION_DENIED / PLATFORM_ROLE_REQUIRED (RequireSuperadmin denial)
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: NOT_FOUND
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Conflict:
      description: CONFLICT (e.g. disabling a mandatory module, revoking the last superadmin)
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    ValidationFailed:
      description: VALIDATION_FAILED, a domain-policy rejection distinct from ordinary field-shape validation
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    AuthorizationUnavailable:
      description: >
        AUTHORIZATION_UNAVAILABLE, authz (or another required dependency) could not be reached;
        fail-closed, never a stale allow.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }